requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==2.2
django-admin-honeypot==1.1.0
django-admin-sortable2==1.0.4
django-adminrestrict==3.0
django-captcha-admin==0.2
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-geoposition==0.3.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-multi-captcha-admin==2.0.0
django-parler==2.3
django-partial-date==1.3.2
django-ranged-response==0.2.0
django-recaptcha==1.3.0
django-recaptcha3==0.4.0
django-rosetta==0.9.8
django-simple-captcha==0.5.17
gunicorn==20.1.0
idna==3.3
pendulum==2.1.2
Pillow==9.0.1
polib==1.1.1
python-dateutil==2.8.2
pytz==2021.3
pytzdata==2020.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
